+++ src/components/menu-category/container/card/Card.tsx
@@ -1,9 +1,8 @@
 import styles from '@/components/menu-category/styles/category.module.css';
-import useSetCategories from '@/hooks/query/menu/menu-category/useSetCategories';
-import useCategoriesStore from '@/shared/store/menu/menu-category';
+import useDragDrop from '@/hooks/service/menu/useDragDrop';
 import { Tables } from '@/types/supabase';
 import clsx from 'clsx';
-import { useEffect, useRef, useState } from 'react';
+import { useEffect } from 'react';
 import EditCategoryComponent from '../form/EditCategory';
 import RemoveCategoryComponent from '../form/RemoveCategory';
 
@@ -14,55 +13,23 @@ interface PropsType {
   setDropNum: React.Dispatch<React.SetStateAction<number>>;
 }
 
-const CategroyCardPage = ({ item, idx, dropNum, setDropNum }: PropsType) => {
-  const { updatePositionMutate } = useSetCategories();
-  const categories = useCategoriesStore(state => state.categories);
-  const [isDragging, setIsDragging] = useState(false);
+const DRAG_TITLE = 'category';
 
-  // 드래그 이벤트
-  const dragItemRef = useRef(0); // 드래그할 아이템의 인덱스
-  const dragOverRef = useRef(0); // 드랍할 위치의 아이템의 인덱스
+const CategroyCardPage = ({ item, idx, dropNum, setDropNum }: PropsType) => {
+  const {
+    isDragging,
+    setIsDragging,
+    dragEnterHandler,
+    dragStartHandler,
+    dropHandler,
+    handleDragLeave,
+    handleDragOver,
+  } = useDragDrop();
 
   useEffect(() => {
     setIsDragging(false);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [idx]);
-
-  // 드래그 시작될 때 실행
-  const dragStartHandler = (e: React.DragEvent<HTMLButtonElement>, index: number) => {
-    dragItemRef.current = index;
-  };
-
-  // 드래그중인 대상이 위로 포개졌을 때
-  const dragEnterHandler = (e: React.DragEvent<HTMLButtonElement>, index: number) => {
-    dragOverRef.current = index;
-    setDropNum(index);
-    setIsDragging(true);
-  };
-
-  // 드래그 중인 요소 위로 이동할 때 스타일 변경
-  const handleDragOver = (e: React.DragEvent<HTMLButtonElement>) => {
-    e.preventDefault();
-    setIsDragging(true);
-  };
-
-  // 드래그 중인 요소가 영역을 떠날 때 스타일 초기화
-  const handleDragLeave = () => {
-    setIsDragging(false);
-  };
-
-  // 드랍 (커서 뗐을 때)
-  const dropHandler = async () => {
-    const newList = [...categories];
-    const dragItemValue = newList[dragItemRef.current];
-    const dragOverValue = newList[dropNum];
-    const dragGroup = {
-      pick: dragItemValue,
-      over: dragOverValue,
-    };
-    updatePositionMutate(dragGroup);
-    dragItemRef.current = 0;
-    dragOverRef.current = 0;
-  };
 
   return (
     <li key={item.id}>
@@ -70,10 +37,10 @@ const CategroyCardPage = ({ item, idx, dropNum, setDropNum }: PropsType) => {
         type="button"
         draggable
         onDragStart={e => dragStartHandler(e, idx)}
-        onDragEnter={e => dragEnterHandler(e, idx)}
+        onDragEnter={e => dragEnterHandler(e, idx, setDropNum)}
         onDragOver={handleDragOver}
         onDragLeave={handleDragLeave}
-        onDragEnd={dropHandler}
+        onDragEnd={() => dropHandler(dropNum, DRAG_TITLE)}
         className={clsx(styles.draggable, {
           [styles.dragging]: isDragging,
         })}
