+++ src/components/menu-category/container/form/FormCategory.tsx
@@ -1,5 +1,6 @@
 import styles from '@/components/menu-category/styles/form.module.css';
 import useSetCategories from '@/hooks/query/menu/menu-category/useSetCategories';
+import useToast from '@/hooks/service/ui/useToast';
 import useCategoriesStore, { setCategory } from '@/shared/store/menu/menu-category';
 import { TablesInsert } from '@/types/supabase';
 import AddCategoryComponent from './AddCategory';
@@ -10,6 +11,7 @@ interface MenuCategoryModal {
 }
 
 const CategoryFormPage: React.FC<MenuCategoryModal> = props => {
+  const { toast } = useToast();
   const isEdit = useCategoriesStore(state => state.isEdit);
   const category = useCategoriesStore(state => state.category);
   const categories = useCategoriesStore(state => state.categories);
@@ -28,6 +30,12 @@ const CategoryFormPage: React.FC<MenuCategoryModal> = props => {
     isEdit ? updateNameMutate(category) : addMutate(newCategoryData);
     setCategory({ ...category, id: '', name: '' });
     props.clickCategoryModalHide();
+    toast(!isEdit ? '카테고리 등록 성공' : '카테고리 수정 성공', {
+      type: 'success',
+      position: 'top-center',
+      showCloseButton: false,
+      autoClose: 2000,
+    });
   };
 
   return (
