+++ src/components/menu-item/items/card/Card.tsx
@@ -1,6 +1,6 @@
 import MenuItemModal from '@/components/menu-item/items/modal/MenuItemModal';
 import styles from '@/components/menu-item/styles/menu-item-card.module.css';
-import useSetMenuItem from '@/hooks/query/menu/menu-item/useSetMenuItems';
+import useDragDrop from '@/hooks/service/menu/useDragDrop';
 import { useModal } from '@/hooks/service/ui/useModal';
 import { convertNumberToWon } from '@/shared/helper';
 import useMenuItemStore, {
@@ -13,7 +13,7 @@ import useMenuOptionStore, { setMenuOption, setMenuOptions } from '@/shared/stor
 import { Tables } from '@/types/supabase';
 import clsx from 'clsx';
 import Image from 'next/image';
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect } from 'react';
 import { FaStar } from 'react-icons/fa';
 import EditButton from '/public/icons/pencil.svg';
 import Default from '/public/whiteLogo.svg';
@@ -25,16 +25,22 @@ interface PropsType {
   setDropNum: React.Dispatch<React.SetStateAction<number>>;
 }
 
+const DRAG_TITLE = 'menu-item';
+
 const MenuItemCard = ({ item, idx, dropNum, setDropNum }: PropsType) => {
   const { MagicModal } = useModal();
-  const { updatePositionMutate } = useSetMenuItem();
   const menuItem = useMenuItemStore(state => state.menuItem);
-  const categoryWithMenuItem = useMenuItemStore(state => state.categoryWithMenuItem);
-  const categoryWithMenuItemList = useMenuItemStore(state => state.categoryWithMenuItemList);
   const menuOption = useMenuOptionStore(state => state.menuOption);
   const origineMenuOptions = useMenuOptionStore(state => state.origineMenuOptions);
-
-  const [isDragging, setIsDragging] = useState(false);
+  const {
+    isDragging,
+    setIsDragging,
+    dragEnterHandler,
+    dragStartHandler,
+    dropHandler,
+    handleDragLeave,
+    handleDragOver,
+  } = useDragDrop();
 
   // 메뉴 선택
   const clickChoiceCategoryHandler = (item: Tables<'menu_item'>) => {
@@ -63,51 +69,10 @@ const MenuItemCard = ({ item, idx, dropNum, setDropNum }: PropsType) => {
     setMenuOptions(filterMenuOptionList);
   };
 
-  // 드래그 이벤트
-  const dragItemRef = useRef(0); // 드래그할 아이템의 인덱스
-  const dragOverRef = useRef(0); // 드랍할 위치의 아이템의 인덱스
-
   useEffect(() => {
     setIsDragging(false);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [idx]);
-
-  // 드래그 시작될 때 실행
-  const dragStartHandler = (e: React.DragEvent<HTMLButtonElement>, index: number) => {
-    dragItemRef.current = index;
-  };
-
-  // 드래그중인 대상이 위로 포개졌을 때
-  const dragEnterHandler = (e: React.DragEvent<HTMLButtonElement>, index: number) => {
-    dragOverRef.current = index;
-    setDropNum(index);
-    setIsDragging(true);
-  };
-
-  // 드래그 중인 요소 위로 이동할 때 스타일 변경
-  const handleDragOver = (e: React.DragEvent<HTMLButtonElement>) => {
-    e.preventDefault();
-    setIsDragging(true);
-  };
-
-  // 드래그 중인 요소가 영역을 떠날 때 스타일 초기화
-  const handleDragLeave = () => {
-    setIsDragging(false);
-  };
-
-  // 드랍 (커서 뗐을 때)
-  const dropHandler = async () => {
-    const filterIndex: number = categoryWithMenuItemList.findIndex(list => list.id === categoryWithMenuItem.id);
-    const newList = [...categoryWithMenuItemList[filterIndex].menu_item];
-    const dragItemValue = newList[dragItemRef.current];
-    const dragOverValue = newList[dropNum];
-    const dragGroup = {
-      pick: dragItemValue,
-      over: dragOverValue,
-    };
-    updatePositionMutate(dragGroup);
-    dragItemRef.current = dropNum;
-    dragOverRef.current = dragItemRef.current;
-  };
 
   return (
     <li
@@ -122,10 +87,10 @@ const MenuItemCard = ({ item, idx, dropNum, setDropNum }: PropsType) => {
         type="button"
         draggable
         onDragStart={e => dragStartHandler(e, idx)}
-        onDragEnter={e => dragEnterHandler(e, idx)}
+        onDragEnter={e => dragEnterHandler(e, idx, setDropNum)}
         onDragOver={handleDragOver}
         onDragLeave={handleDragLeave}
-        onDragEnd={dropHandler}
+        onDragEnd={() => dropHandler(dropNum, DRAG_TITLE)}
         onClick={() => clickChoiceCategoryHandler(item)}
         className={clsx(styles.draggable, {
           [styles.dragging]: isDragging,
